requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
google-generativeai==0.8.5
pytz
telethon
